check_tidymass_version

Check if there are packages in tidymass can be updated.

Usage

check_tidymass_version(packages = c("core", "all"))

Arguments

packages

core or all packages in tidymass. "core" means all the core packages in tidymass and "all" means all the packages in tidymass.

Author

Xiaotao Shen shenxt1990@outlook.com