Conflicts between the tidymass and other packages

This function lists all the conflicts between packages in the tidymass and other packages that you have loaded.

tidymass_conflicts()

Details

There are four conflicts that are deliberately ignored: intersect, union, setequal, and setdiff from dplyr. These functions make the base equivalents generic, so shouldn't negatively affect any existing code.

Examples

tidymass_conflicts()
#> ── Conflicts ─────────────────────────────────────────── tidymass_conflicts() ──
#> ✖ ComplexHeatmap::Heatmap() masks massstat::Heatmap()
#> ✖ massdataset::apply()      masks base::apply()
#> ✖ BiocGenerics::colMeans()  masks massdataset::colMeans(), base::colMeans()
#> ✖ BiocGenerics::colSums()   masks massdataset::colSums(), base::colSums()
#> ✖ xcms::collect()           masks dplyr::collect()
#> ✖ MSnbase::combine()        masks Biobase::combine(), BiocGenerics::combine(), dplyr::combine()
#> ✖ tidyr::extract()          masks magrittr::extract()
#> ✖ metid::filter()           masks metpath::filter(), tidygraph::filter(), dplyr::filter(), massdataset::filter(), stats::filter()
#> ✖ S4Vectors::first()        masks dplyr::first()
#> ✖ tidygraph::groups()       masks xcms::groups(), dplyr::groups()
#> ✖ S4Vectors::head()         masks utils::head()
#> ✖ S4Vectors::intersect()    masks BiocGenerics::intersect(), massdataset::intersect(), base::intersect()
#> ✖ dplyr::lag()              masks stats::lag()
#> ✖ mixOmics::pls()           masks massstat::pls()
#> ✖ mixOmics::plsda()         masks massstat::plsda()
#> ✖ Rcpp::prompt()            masks utils::prompt()
#> ✖ BiocGenerics::relist()    masks utils::relist()
#> ✖ tidygraph::rename()       masks S4Vectors::rename(), dplyr::rename(), massdataset::rename()
#> ✖ BiocGenerics::rowMeans()  masks massdataset::rowMeans(), base::rowMeans()
#> ✖ BiocGenerics::rowSums()   masks massdataset::rowSums(), base::rowSums()
#> ✖ tidygraph::select()       masks MASS::select(), dplyr::select(), massdataset::select()
#> ✖ S4Vectors::stack()        masks utils::stack()
#> ✖ S4Vectors::tail()         masks utils::tail()